feat($search): add posts filter in category and tag pages (#18)
